Hide list view group headers for groups with an empty title
This amends list view (including the playlist view) grouping logic to not show group headers when they have an empty string as a label.

There is no change to how grouping itself is performed. This means, for example, items are consistently indented according to the total number of grouping levels (if indentation is enabled).

To force a group header with an empty label to be shown, use a zero-width space (`$char(8203)`).