Set Buttons toolbar separator size explicitly in light mode
This sets the width of separators explicitly in the Buttons toolbar in light mode. This resolves a problem where separators had a a fair bit of padding to the right of them at low DPIs.

(It turns out, for separators, the `iBitmap` member of `TBBUTTON` is the width in pixels of the separator. But the toolbar control does not centre the separator line in the space occupied, nor does it DPI-scale the separator line, so this is all still a bit messy.)

For dark mode, native toolbar separators are not being used as they did not render correctly when last tested.

When switching between light and dark mode, the panel host is now told that the minimum width has changed, as there can still be slight differences in separator padding between light and dark mode.