Don’t reload playlist view artwork on non-metadata changes
This makes the playlist view use `metadb_io_callback::on_changed_sorted()` (instead of `playlist_callback::on_items_modified()`) for triggering artwork flushing on metadata changes.

Changes where the `p_fromhook` argument is true are ignored, as these are due to things like custom title-formatting fields provided by other components changing and are usually irrelevant.

(The old behaviour was, for example, causing playlist view artwork to reload for the playing or focused track when the volume was changed with the Output Info component.)