Avoid unwanted playlist updates when removing Filter search toolbar
If a query was entered in the Filter search toolbar and autosend was enabled in Preferences, removing the toolbar from the layout would always update the Filter results playlist (with the entire media library contents).

This stops the toolbar doing that when there are no Filter panels in the layout. It also avoids doing it if there are Filter panels but they are also being removed (for example, during a layout preset switch). (Avoiding it in the latter scenario is a best effort, as it’s not a straightforward scenario to detect.)

If the toolbar is removed from the layout but Filter panels remain, they will be updated as normal, and autosend will be triggered if enabled in Preferences.