Add an option to make playlist view artwork sticky
This adds an option for playlist view artwork to be sticky, so that it stays in view when scrolling down while the group is still on screen.

Some changes were made in the underlying list view component as part of this which should also make rendering more efficient in some circumstances.

A bug where the drag-and-drop insertion marker was rendered in the wrong place when the list view was scrolled horizontally was also fixed.

Additionally, playlist view artwork settings are now written to FCL files.