Cache DirectWrite text layouts
This updates ui_helpers to make use of an LRU cache for DirectWrite text layouts to improve performance when the same text is rendered multiple times with the same maxium width and height.

This uses a bit more memory (around 5 to 10 MB or so). Cache size limits are set a bit lower on 32-bit builds, and caching is disabled if the foobar2000 low memory mode is active.