Reduce Direct2D panels memory consumption if resized while hidden
This resolves a problem where Item details and Artwork view temporarily consumed a lot of memory if they were resized while hidden (for example, when in an inactive tab).

Presumably resources were building up which were then not flushed until the panel became visible again.

This just stops resizing the render target in WM_SIZE, and instead adds a check at the beginning of WM_PAINT to see if the render target needs resizing. This solves the problem and seems to work well enough.