Optimise spectrum analyser rendering
This makes the spectrum analyser use a device-independent bitmap (DIB) and modify the bitmap bits directly, rather than use drawing operations as previously.

This makes it a fair bit more efficient, and was the most efficient rendering method that was found.

Support for the legacy `uie::visualisation` service was also removed, and the code has been simplified as a result.