Enable sorting in Filter search
This enables sorting in queries entered in the Filter search toolbar.

Explicit sorting using SORT BY is now always used and respected.

Implicit sorting (for example, for time-related queries) is used only if the sorting pattern in Filter preferences is disabled.

The dodgy wording of the sorting option in Filter preferences was also improved.