Fix incorrect handling of colour codes for selections
This resolves a problem where colour codes (`$rgb()`) did not behave correctly once the relevant item was selected.

This was related to caching logic added for DirectWrite text layouts.

This amends how colour codes are handled so that the non-selection and selection colours are selected between at render time, rather than when the colour code is processed, to fix the problem without defeating the caching.