Work around inconsistent window occlusion status when display is off
This works around a problem where Direct2D and DXGI cannot decide if a window is occluded when the display is turned off.

For `ID2D1HwndRenderTarget`, the `CheckWindowState` method reports the window is occluded after `EndDraw` is called, but then when called again it reports that the window isn’t occluded.

For `IDXGISwapChain`, it seems like test presentations report that the window isn’t occluded, while real ones report that it is occluded.

For Item details, the inconsistency was causing it to think that occlusion status was continuously changing, leading to excessive drawing activity.

This change amends the logic in Item details so that it doesn’t trust `ID2D1HwndRenderTarget::CheckWindowState()` reporting that a window isn’t occluded so much when it isn’t called immediately after `ID2D1HwndRenderTarget::EndDraw()`.