Add option to disable extra playlist view artwork vertical spacing
This adds an option to control whether extra playlist view spacing below leaf group headers and after the last item of a group is added when artwork is enabled.

The option is named ‘Add extra vertical spacing to increase distance between group headers and artwork’ and is on the Artwork tab on the Playlist view preferences page and is enabled by default. When disabled, the extra vertical spacing is removed, restoring similar vertical spacing as versions prior to 3.2.0-beta.1.

Additionally, the ui_helpers update fixes a problem where, when the extra spacing is enabled, it was not present for the first group. A minor error in hit-testing calculations was also fixed.