Preserve selection when right-clicking selected group
This amends the logic when right-clicking on a group header in the playlist view, and in Item properties, so that if all the items in the group are already selected, the existing selection is not modified.

This is more consistent with the behaviour when right-clicking items.

There is no change to the behaviour if all items in the group are not selected. In this case, the group is exclusively selected as before.