Don’t use front cover stub image in Artwork view for other image types
This stops the Artwork view using the front cover stub image configured on the Display preferences page as a fallback for other stub image types (which was the behaviour added in #345).

Stub images can be explicitly configured for each image type, so using the front cover stub image as a fallback for other stub image types doesn’t make much sense. The path of the front cover stub image can be copied to the other stub image path fields if needed.

The built-in placeholder image is still used when no stub image is defined in Preferences for the artwork type.