Update %_display_index% in playlist view after item edits
This fixes a bug where styles scripts were not re-evaluated when item modifications caused the `%_display_index%` values for subsequent items to change. This would happen if an edit to an item caused group headers to be shown or removed. This broke the shading of alternate rows, for example.

Additionally, some changes were made to make the logic around whether a group is shown or hidden more consistent. In theory the changes shouldn’t make a difference. The old logic relied on parent groups being different if a child group is different, which should be the case. However, a crash was submitted indicating there may be another bug lurking somewhere breaking that assumption. So far, I haven’t been able to find a scenario that does break that assumption, though.